feat : api/game/games api 생성 중, Error 처리 진행 중.

+++ back/src/main/java/bunsan/returnz/domain/game/service/GameStockPriceInformationDto.java
@@ -14,8 +14,8 @@ import lombok.Setter;
 @AllArgsConstructor
 public class GameStockPriceInformationDto {
 	Integer historyDate;
-	Integer historyPrice;
-	Integer historyDiff;
+	Double historyPrice;
+	Double historyDiff;
 	StockState historyUpAndDown;
 	Long Volume;
 }
